fix: update PR #36 changes
